refactor: extract docx section helpers
